refactor: introduce spec and policy packages; re-export types for compatibility
- Move core types/interfaces to spec package\n- Add policy/failure and policy/routing packages\n- Re-export types and policies at root for backward compatibility\n- Update README and AGENTS docs\n\nNo public API break; existing imports continue to work.
