+++ spec/doc.go
@@ -1,0 +1,4 @@
+// Package spec defines the public API contract for sqsrouter.
+// It holds core data structures and interfaces shared across the router
+// and built-in policy implementations.
+package spec
